requirements.txt
torch
transformers
peft
datasets==3.6.0
trl
bitsandbytes
sentence-transformers
faiss-cpu
scikit-learn
accelerate
sacrebleu
pymupdf
easyocr
langchain
tqdm
langchain-text-splitters
numpy
kokoro>=0.9.2
soundfile
pydub
jupyterlab
